+++ src/ArduinoJson/Object/ObjectRef.hpp
@@ -213,7 +213,7 @@ class ObjectRef : public ObjectRefBase<CollectionData>,
   FORCE_INLINE void remove(iterator it) const {
     if (!_data)
       return;
-    _data->remove(it.internal());
+    _data->removeSlot(it.internal());
   }
 
   // remove(const std::string&) const